[WIP] Optimize usability for Time ranges. (#17)
* Change the model to support serializer.
After serialization using Gson, include all the data into one tag to provide data acquisition capability for the new javascript-based frontend.

* Edit readme.

* Add pictures.

* Demo completed.

* Create Management Link.

* Add Request Handler

* Delete generate js chunks.

* Add getExcludedDates handler.

* Implement get/set excluded-dates function.

* Integrate react project.
Remove build chunks.

* Jenkins Style UI.

* Add a new Readme

* Add proxy.

* Add axios and api.

* Fetching on open/Save on close.

* Optimize display on excluded dates.

* Add time ranges.

* Add time ranges.

* Add type and utcOffset field.

* Add preset type.

* Date Type in progress.

* Add Comments about selecting presets.

* Apply regional days works with Chinese New Year algorithm.

* Add algorithm for chinese lunar date.

* Add algorithm for chinese lunar date.

* Add directories for excluded date and time ranges.

Move getBrief() methods to utils.js.

Use range function instead of plain repeats in select options.

Change to partial import for lodash.

* Set default value of date to "". Update intro and copyright.

* Add api for time ranges.

* Alter directory structure.

* Add timezone select.

* Time Range Api.

* Add workflow doc.

* Extract form items.

* Disable java tests.(Will implement later)

* Use timezone select.

* Time range input validation.

* Add time range slider.

* Customize time range slider.

* Update .gitignore to ignore node built files and node_modules.

* Modify class to fix findbugs error.

* Disable tests.

* Sort time ranges by day of week.

* Add .editorconfig

* Reset components structure with adding timeRangeContainer and excludedDateContainer.

* Add debounced saving function/Add saving/loading indicator for time ranges.

* Add comments.

* Update function name.

* Add webpack config file for test.

* UI Optimize

* Formats.
