+++ src/main/react/app/containers/App/constants.js
@@ -1,13 +1,13 @@
 import { fetchTimezones } from "../../api";
 
 export const WEEKDAYS = {
-  Sunday: 1,
-  Monday: 2,
-  Tuesday: 3,
-  Wednesday: 4,
-  Thursday: 5,
-  Friday: 6,
-  Saturday: 7
+  Sunday: 0,
+  Monday: 1,
+  Tuesday: 2,
+  Wednesday: 3,
+  Thursday: 4,
+  Friday: 5,
+  Saturday: 6
 };
 
 export const PERIODS = {
