+++ src/main/react/package.json
@@ -19,6 +19,7 @@
     "build": "cross-env NODE_ENV=production webpack --config internals/webpack/webpack.prod.babel.js --color -p --progress --hide-modules --display-optimization-bailout",
     "build:clean": "rimraf ./build",
     "start": "cross-env NODE_ENV=development node server",
+    "start:watch-build": "cross-env NODE_ENV=development webpack --config internals/webpack/webpack.prod.babel.js --color -p --progress --hide-modules --display-optimization-bailout --watch",
     "start:tunnel": "cross-env NODE_ENV=development ENABLE_TUNNEL=true node server",
     "start:production": "npm run test && npm run build && npm run start:prod",
     "start:prod": "cross-env NODE_ENV=production node server",
@@ -63,8 +64,6 @@
     "connected-react-router": "4.5.0",
     "cross-env": "5.2.0",
     "express": "4.16.4",
-    "fontfaceobserver": "2.0.13",
-    "history": "4.7.2",
     "hoist-non-react-statics": "3.0.1",
     "immutable": "3.8.2",
     "intl": "1.2.5",
@@ -75,6 +74,7 @@
     "minimist": "1.2.0",
     "moment": "^2.22.1",
     "moment-timezone": "^0.5.25",
+    "only": "^0.0.2",
     "prop-types": "15.6.2",
     "qs": "latest",
     "rc-slider": "^8.6.13",
@@ -115,6 +115,7 @@
     "circular-dependency-plugin": "5.0.2",
     "compare-versions": "3.4.0",
     "compression-webpack-plugin": "2.0.0",
+    "copy-webpack-plugin": "^5.0.3",
     "coveralls": "3.0.2",
     "css-loader": "1.0.0",
     "enzyme": "3.7.0",
